chore: move details to container
